requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-filter==22.1
django-search-listview==0.2.1
django-widget-tweaks==1.4.12
mysqlclient==2.1.1
numpy==1.23.2
pandas==1.4.3
Pillow==9.4.0
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
